Throwing Overhaul IV: Returning and Penetration
To differentiate the different types of throwables, javelins now always
penetrate. This overlaps some with crossbows, which provide a
penetrating brand, but a crossbow user can choose from many brands and
has access to weapon enchantment.

Tomahawks always return. This is more a QoL change than anything, and a
step towards a broader ranged reform. In terms of ammo availability, the
mulch rate remains the same, so throws per tomahawk at on strategic
timescales are static. On tactical timescales, a player relying on
tomahawks for damage will for the most part not be limited by ammo; a
player 0-skill throwing tomahawks will get a few extra throws in the
early game, but the damage from these extra throws isn't that much.
Since this effect applies to monsters also, giving them a comparative
buff, it should be a wash.